pyproject.toml
[tool.poetry]
name = "python-assessment"
version = "0.1.0"
description = "Assessment for the pipeline talentql program"
authors = ["David Ajawu <ajawudavid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
